+++ packages/extension-base/src/background/RequestBytesSign.ts
@@ -3,9 +3,11 @@
 
 import type { KeyringPair } from '@polkadot/keyring/types';
 import type { SignerPayloadRaw } from '@polkadot/types/types';
+import type { HexString } from '@polkadot/util/types';
 import type { RequestSign } from './types';
 
-import { wrapBytes } from '@polkadot/extension-dapp/wrapBytes';
+import { wrapBytes } from '@reef-defi/extension-dapp/wrapBytes';
+
 import { TypeRegistry } from '@polkadot/types';
 import { u8aToHex } from '@polkadot/util';
 
@@ -16,7 +18,7 @@ export default class RequestBytesSign implements RequestSign {
     this.payload = payload;
   }
 
-  sign (_registry: TypeRegistry, pair: KeyringPair): { signature: string } {
+  sign (_registry: TypeRegistry, pair: KeyringPair): { signature: HexString } {
     return {
       signature: u8aToHex(
         pair.sign(