+++ packages/extension-compat-metamask/src/index.ts
@@ -1,8 +1,9 @@
 // Copyright 2019-2020 @polkadot/extension-dapp authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { Injected, InjectedAccount, InjectedWindow } from '@polkadot/extension-inject/types';
+import type { Injected, InjectedAccount, InjectedWindow } from '@reef-defi/extension-inject/types';
 import type { SignerPayloadRaw, SignerResult } from '@polkadot/types/types';
+import type { HexString } from '@polkadot/util/types';
 
 import detectEthereumProvider from '@metamask/detect-provider';
 import Web3 from 'web3';
@@ -74,7 +75,7 @@ function injectMetaMaskWeb3 (win: Web3Window): void {
         },
         signer: {
           signRaw: async (raw: SignerPayloadRaw): Promise<SignerResult> => {
-            const signature = (await provider.request({ method: 'eth_sign', params: [raw.address, Web3.utils.sha3(raw.data)] })) as string;
+            const signature = (await provider.request({ method: 'eth_sign', params: [raw.address, Web3.utils.sha3(raw.data)] })) as HexString;
 
             return { id: 0, signature };
           }