+++ packages/extension-ui/src/components/translate.ts
@@ -3,7 +3,7 @@
 
 import { useTranslation as useTranslationBase, UseTranslationResponse, withTranslation } from 'react-i18next';
 
-export function useTranslation (): UseTranslationResponse<string> {
+export function useTranslation (): UseTranslationResponse<string, string> {
   return useTranslationBase();
 }